+++ cucumber-tests/src/test/java/org/eclipse/tractusx/irs/cucumber/E2ETestStepDefinitions.java
@@ -151,7 +151,7 @@ public class E2ETestStepDefinitions {
                                    .get("/irs/jobs/" + jobId)
                                    .as(Jobs.class)
                                    .getJob()
-                                   .getJobState()
+                                   .getState()
                                    .equals(JobState.value(status)));
 
         completedJob = given().spec(authProperties.getNewAuthenticationRequestSpecification())