+++ irs-models/src/main/java/org/eclipse/tractusx/irs/component/JobStatusResult.java
@@ -36,17 +36,18 @@ import org.eclipse.tractusx.irs.component.enums.JobState;
 @Value
 @Jacksonized
 @Builder(toBuilder = true)
+@SuppressWarnings("PMD.ShortVariable")
 public class JobStatusResult {
 
     /**
      * Job identifier
      */
-    private UUID jobId;
+    private UUID id;
 
     /**
      * Current state of the job
      */
-    private JobState jobState;
+    private JobState state;
 
     /**
      * Timestamp when the job was started
@@ -58,6 +59,6 @@ public class JobStatusResult {
      * Timestamp when the job was completed
      */
     @Schema(implementation = ZonedDateTime.class)
-    private ZonedDateTime jobCompleted;
+    private ZonedDateTime completedOn;
 
 }